server/pyproject.toml
[project]
name = "mozi-server"
version = "0.1.0"
description = "Mozi backend — microservices for auth, workspace, agent, sandbox, MCP"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.15",
    "pydantic>=2.10",
    "email-validator>=2.1",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.28",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.3",
    "pgvector>=0.3",
    "minio>=7.2",
    "structlog>=24.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["shared", "services"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
